refactor: 重构项目架构，移除旧状态系统并统一使用三层状态机
- 删除旧架构文件(fd2_game.h, fd2_states.h等)
- 更新main.c使用新状态机架构
- 完善全局变量系统
- 实现场景注册系统fd2_register_all_scenes()
- 更新Makefile移除旧文件引用
- 添加输入扫描码系统框架
- 重构资源加载系统
- 完善存档系统实现

+++ src/fd2_icon_b24.c
@@ -145,7 +145,7 @@ int fd2_icon_get(int icon_id) {
         return -1;
     }
 
-    if (icon_id < 0 || icon_id >= g_total_icons) {
+    if (icon_id < 0 || (unsigned int)icon_id >= (unsigned int)g_total_icons) {
         fprintf(stderr, "fd2_icon_get: icon_id %d out of range (max %d)\n",
                 icon_id, g_total_icons);
         return -1;
